refactor(triple): remove io to new mod
